Remove internal trace reporting types.
