+++ trace/spanstore.go
@@ -86,17 +86,9 @@ func (_ internalOnly) ReportSpansByError(name string, code int32) []*SpanData {
 	return out
 }
 
-// BucketConfiguration stores the number of samples to store for span buckets
-// for successful and failed spans for a particular span name.
-type BucketConfiguration struct {
-	Name                 string
-	MaxRequestsSucceeded int
-	MaxRequestsErrors    int
-}
-
 // ConfigureBucketSizes sets the number of spans to keep per latency and error
 // bucket for different span names.
-func (_ internalOnly) ConfigureBucketSizes(bcs []BucketConfiguration) {
+func (_ internalOnly) ConfigureBucketSizes(bcs []internal.BucketConfiguration) {
 	for _, bc := range bcs {
 		latencyBucketSize := bc.MaxRequestsSucceeded
 		if latencyBucketSize < 0 {
@@ -117,24 +109,24 @@ func (_ internalOnly) ConfigureBucketSizes(bcs []BucketConfiguration) {
 }
 
 // ReportSpansPerMethod returns a summary of what spans are being stored for each span name.
-func (_ internalOnly) ReportSpansPerMethod() map[string]PerMethodSummary {
-	out := make(map[string]PerMethodSummary)
+func (_ internalOnly) ReportSpansPerMethod() map[string]internal.PerMethodSummary {
+	out := make(map[string]internal.PerMethodSummary)
 	ssmu.RLock()
 	defer ssmu.RUnlock()
 	for name, s := range spanStores {
 		s.mu.Lock()
-		p := PerMethodSummary{
+		p := internal.PerMethodSummary{
 			Active: len(s.active),
 		}
 		for code, b := range s.errors {
-			p.ErrorBuckets = append(p.ErrorBuckets, ErrorBucketSummary{
+			p.ErrorBuckets = append(p.ErrorBuckets, internal.ErrorBucketSummary{
 				ErrorCode: code,
 				Size:      b.size(),
 			})
 		}
 		for i, b := range s.latency {
 			min, max := latencyBucketBounds(i)
-			p.LatencyBuckets = append(p.LatencyBuckets, LatencyBucketSummary{
+			p.LatencyBuckets = append(p.LatencyBuckets, internal.LatencyBucketSummary{
 				MinLatency: min,
 				MaxLatency: max,
 				Size:       b.size(),
@@ -183,25 +175,6 @@ func (_ internalOnly) ReportSpansByLatency(name string, minLatency, maxLatency t
 		}
 	}
 	return out
-}
-
-// PerMethodSummary is a summary of the spans stored for a single span name.
-type PerMethodSummary struct {
-	Active         int
-	LatencyBuckets []LatencyBucketSummary
-	ErrorBuckets   []ErrorBucketSummary
-}
-
-// LatencyBucketSummary is a summary of a latency bucket.
-type LatencyBucketSummary struct {
-	MinLatency, MaxLatency time.Duration
-	Size                   int
-}
-
-// ErrorBucketSummary is a summary of an error bucket.
-type ErrorBucketSummary struct {
-	ErrorCode int32
-	Size      int
 }
 
 // spanStore keeps track of spans stored for a particular span name.
